requirements.txt
pyshorteners
pyperclip
qrcode
Pillow
tinydb
